remove repos.yml.default
Originally it was planned to have repos.yml not in the repo and
have KBS copy it from repos.yml.default.
That never happened as repos.yml got checked in as part of a "TEMP" commit
and was then forgotten about.

Additionally, the files have started to drift.

New plan: read repos.local.yml if available. Added it to gitignore!